Revert QueryParameters to map[string]string and add QueryParametersAny
This keeps source, binary and runtime behavior in a backward compatible
way. It does however require to do some bookkeeping twice. The old
QueryParameters is still available and works in the same way. Most
values are still stored in the old QueryParameters. Only lists are also
stored in the new QueryParametersAny. Any value in QueryParametersAny
will override the value from QueryParameters.